# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dushi Stores Api Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "info@dushistores.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "3.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/subcategories": {
            "get": {
                "tags": [
                    "SubCategories"
                ],
                "summary": "Get All SubCategories",
                "description": "get all subcategories",
                "operationId": "7328e073e16bef20d03b6e9650f37f3c",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/api/subcategories/{subCategory_id}": {
            "get": {
                "tags": [
                    "SubCategories"
                ],
                "summary": "Show SubCategory",
                "description": "show subcategory",
                "operationId": "b457c52f8819cc6a8b2d3ee353f9bd07",
                "parameters": [
                    {
                        "name": "subCategory_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/api/wishlist": {
            "get": {
                "tags": [
                    "Wishlist"
                ],
                "summary": "Get User Wishlist",
                "description": "get user wishlist",
                "operationId": "a10ee44a9d61babdc7874c7624617e44",
                "parameters": [
                    {
                        "name": "X-USER-KEY",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/api/wishlist/add/{product_id}": {
            "post": {
                "tags": [
                    "Wishlist"
                ],
                "summary": "Add To Wishlist",
                "description": "add to wishlist",
                "operationId": "fb0ae06048c583d49f2cfaf198ff3120",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "X-USER-KEY",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/api/wishlist/remove/{product_id}": {
            "delete": {
                "tags": [
                    "Wishlist"
                ],
                "summary": "Remove From Wishlist",
                "description": "remove from wishlist",
                "operationId": "7d1f496a2167768532a8bf8cd403c0df",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "X-USER-KEY",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "description": "Enter token",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "bearer_token": []
        }
    ]
}